Moved Services headers from include to src
PrintLoadingPlugins and Timing are not Services with which users should interact. Therefore their headers do not need to be included in any code outside of their package.
