+++ FWCore/Services/plugins/Module.cc
@@ -4,13 +4,13 @@
 #include "FWCore/Services/src/UnixSignalService.h"
 
 #include "FWCore/Services/src/JobReportService.h"
-#include "FWCore/Services/interface/Timing.h"
+#include "FWCore/Services/src/Timing.h"
 #include "FWCore/Services/src/Memory.h"
 #include "FWCore/Services/src/CPU.h"
 #include "FWCore/Services/src/LoadAllDictionaries.h"
 #include "FWCore/Services/src/EnableFloatingPointExceptions.h"
 #include "FWCore/ServiceRegistry/interface/ServiceMaker.h"
-#include "FWCore/Services/interface/PrintLoadingPlugins.h"
+#include "FWCore/Services/src/PrintLoadingPlugins.h"
 
 using edm::service::JobReportService;
 using edm::service::Tracer;
